Refactor project structure to N-Tier architecture
