+++ ProjectBankAPI.Application/Queries/Transactions/GetAllTransactionsHandler.cs
@@ -1,0 +1,24 @@
+﻿using MediatR;
+using ProjectBankAPI.Infrastructure.Persistence.Repositories;
+using ProjectBankAPI.Domain.Models;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace ProjectBankAPI.Queries.Transactions
+{
+    public class GetAllTransactionsHandler : IRequestHandler<GetAllTransactionsQuery, IEnumerable<Transaction>>
+    {
+        private readonly ITransactionRepository _repository;
+
+        public GetAllTransactionsHandler(ITransactionRepository repository)
+        {
+            _repository = repository;
+        }
+
+        public async Task<IEnumerable<Transaction>> Handle(GetAllTransactionsQuery request, CancellationToken cancellationToken)
+        {
+            return await _repository.GetAllAsync();
+        }
+    }
+}
