+++ ProjectBankAPI.Domain/Models/BankAccount.cs
@@ -1,0 +1,16 @@
+﻿using System.Text.Json.Serialization;
+
+namespace ProjectBankAPI.Domain.Models
+{
+    public class BankAccount
+    {
+        public int Id { get; set; }
+        public int ClientId { get; set; }
+
+        [JsonIgnore]
+        public Client? Client { get; set; } // Permitir valores nulos
+        public string AccountNumber { get; set; } = string.Empty;
+        public decimal Balance { get; set; }
+        public List<Transaction> Transactions { get; set; } = new();
+    }
+}
