+++ ProjectBankAPI.sln
@@ -7,6 +7,12 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProjectBankAPI", "ProjectBa
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnitTestingXProjectBank.Tests", "UnitTestingXProjectBank.Tests\UnitTestingXProjectBank.Tests.csproj", "{F1FB8084-ABBB-4E1C-8173-907FA03E35A1}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProjectBankAPI.Application", "ProjectBankAPI.Application\ProjectBankAPI.Application.csproj", "{B02C96F5-76F4-4209-BCAB-3C667A94C35D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProjectBankAPI.Domain", "ProjectBankAPI.Domain\ProjectBankAPI.Domain.csproj", "{1E3CA0B1-6032-4F38-85B0-CDFEF131BCF4}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProjectBankAPI.Infrastructure", "ProjectBankAPI.Infrastructure\ProjectBankAPI.Infrastructure.csproj", "{146119F9-ACC6-4129-A59F-27DE38C2A233}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +27,18 @@ Global
 		{F1FB8084-ABBB-4E1C-8173-907FA03E35A1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F1FB8084-ABBB-4E1C-8173-907FA03E35A1}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F1FB8084-ABBB-4E1C-8173-907FA03E35A1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B02C96F5-76F4-4209-BCAB-3C667A94C35D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B02C96F5-76F4-4209-BCAB-3C667A94C35D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B02C96F5-76F4-4209-BCAB-3C667A94C35D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B02C96F5-76F4-4209-BCAB-3C667A94C35D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1E3CA0B1-6032-4F38-85B0-CDFEF131BCF4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1E3CA0B1-6032-4F38-85B0-CDFEF131BCF4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1E3CA0B1-6032-4F38-85B0-CDFEF131BCF4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1E3CA0B1-6032-4F38-85B0-CDFEF131BCF4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{146119F9-ACC6-4129-A59F-27DE38C2A233}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{146119F9-ACC6-4129-A59F-27DE38C2A233}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{146119F9-ACC6-4129-A59F-27DE38C2A233}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{146119F9-ACC6-4129-A59F-27DE38C2A233}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
