--- ProjectBankAPI/Queries/Transactions/GetAllTransactionsQuery.cs
@@ -1,8 +1,0 @@
-﻿using MediatR;
-using ProjectBankAPI.Models;
-using System.Collections.Generic;
-
-namespace ProjectBankAPI.Queries.Transactions
-{
-    public class GetAllTransactionsQuery : IRequest<IEnumerable<Transaction>> { }
-}
