--- ProjectBankAPI/WeatherForecast.cs
@@ -1,13 +1,0 @@
-namespace ProjectBankAPI
-{
-    public class WeatherForecast
-    {
-        public DateOnly Date { get; set; }
-
-        public int TemperatureC { get; set; }
-
-        public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
-
-        public string? Summary { get; set; }
-    }
-}
